Fixed more ranking issues in the main leaderboad + other additions
Added more code to the point system - is till not complete or working.

Ranking seems better, with new code logic. Cleaned up some code to add to the leaderboard. It now compiles the information in the text boxes and sends it to a method designed to write to the data grid.

+++ TASCompDisplay/Scores.cs
@@ -10,9 +10,9 @@ namespace TASCompDisplay
 	{
 		public int Rank { get; set; }
 		public string Username { get; set; }
-		public int Score { get; set; }
+		public double Score { get; set; }
 
-		public Scores(int rank, string username, int score)
+		public Scores(int rank, string username, double score)
 		{
 			Rank = rank;
 			Username = username;
